pyproject.toml
[tool.poetry]
name = "my-fastapi-app"
version = "0.1.0"
description = "A FastAPI application"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.22.0"
opencv-python-headless = "^4.5.5.64"
pillow = "^10.3.0"
tqdm = "^4.66.4"
psutil = "^6.0.0"
matplotlib = "^3.9.0"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
